+++ api/server.js
@@ -2,6 +2,7 @@ const Glue = require('glue');
 const ioc = require('electrolyte');
 const handlebars = require('handlebars');
 const manifestPromise = require('./manifest');
+const bookshelf = require('./src/lib/bookshelf')();
 
 var options = {
   relativeTo: __dirname + '/src'
@@ -16,6 +17,25 @@ manifestPromise.then(manifest => {
     server.auth.strategy('access_token', 'access_token', { token_type: 'access_token' });
     server.auth.strategy('client_credentials', 'access_token', { token_type: 'client_credentials' });
     server.auth.strategy('oidc_session', 'oidc_session');
+    server.auth.strategy('email_token', 'email_token', {
+      findToken: async id => {
+        let token = await bookshelf.model('email_token')
+          .forge({ token: id })
+          .where('expires_at', '>', bookshelf.knex.fn.now())
+          .fetch();
+
+        if (!token) {
+          token = await bookshelf.model('user_password_reset_token')
+            .forge({ token })
+            .where('expires_at', '>', bookshelf.knex.fn.now())
+            .fetch();
+        }
+        return token;
+      },
+      findUser: async(id) => {
+        return await bookshelf.model('user').where({ id }).fetch();
+      }
+    });
 
     ioc.use(id => {
       if (id === 'server') {