+++ api/src/application/client/client-service.js
@@ -96,10 +96,6 @@ module.exports = (bookshelf) => ({
       .fetch({ withRelated: oidcRelations });
   },
 
-  findByRedirectUriAndClientId(clientId, redirect_uri) {
-    return bookshelf.model('client_redirect_uri').where({ client_id: clientId, uri: redirect_uri }).fetchAll();
-  },
-
   update(id, payload) {
     const toStore = Object.assign({}, payload);
     return bookshelf.model('client').forge({ client_id: id }).save(toStore);