+++ api/src/plugins/oidc-session-scheme.js
@@ -5,17 +5,23 @@ exports.register = function (server, pluginOptions, next) {
     return {
 
       async authenticate(request, reply) {
-        const session = await request.server.plugins['open-id-connect'].provider.Session.find(request.state._session);
-
-        if (!session.accountId()) {
+        if (!request.state._session) {
           reply(Boom.forbidden());
         } else {
-          reply.continue({
-            credentials: session,
-          });
+          try {
+            var session = await request.server.plugins['open-id-connect'].provider.Session.find(request.state._session);
+            if (!session.accountId()) {
+              reply(Boom.forbidden());
+            } else {
+              reply.continue({
+                credentials: session,
+              });
+            }
+          } catch(e) {
+            reply(e);
+          }
         }
       }
-
     };
   });